Fix a lot of compiler warnings, disable many
Fix compiler warnings in debug build

qboolean->bool

+++ src/game/g_admin.c
@@ -2328,7 +2328,7 @@ qboolean G_admin_listlayouts( gentity_t *ent )
   int count = 0;
   char *s;
   char layout[ MAX_QPATH ] = { "" };
-  int i = 0;
+  size_t i = 0;
 
   if( trap_Argc( ) == 2 )
     trap_Argv( 1, map, sizeof( map ) );
@@ -2465,7 +2465,7 @@ static qboolean ban_matchname( void *ban, const void *name )
 }
 static void ban_out( void *ban, char *str )
 {
-  int i;
+  size_t i;
   int colorlen1 = 0;
   char duration[ MAX_DURATION_LENGTH ];
   char *d_color = S_COLOR_WHITE;
@@ -2546,7 +2546,7 @@ qboolean G_admin_adminhelp( gentity_t *ent )
   g_admin_command_t *c;
   if( trap_Argc() < 2 )
   {
-    int i;
+    size_t i;
     int count = 0;
 
     ADMBP_begin();
